ecommerceapp/requirements.txt
amqp==5.2.0
asgiref==3.7.2
attrs==23.2.0
billiard==4.2.0
celery==5.3.6
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
coreapi==2.3.3
coreschema==0.0.4
cron_descriptor==1.4.0
cryptography==41.0.7
defusedxml==0.7.1
dj-rest-auth==5.0.2
Django==4.2.9
django-allauth==0.60.1
django-celery-beat==2.5.0
django-resized==1.0.2
django-rest-swagger==2.2.0
django-timezone-field==6.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.0
drf-yasg==1.21.7
flake8==7.0.0
gunicorn==21.2.0
idna==3.6
inflection==0.5.1
iniconfig==2.0.0
itypes==1.2.0
Jinja2==3.1.3
jsonschema==4.21.0
jsonschema-specifications==2023.12.1
kombu==5.3.5
MarkupSafe==2.1.3
mccabe==0.7.0
oauthlib==3.2.2
openapi-codec==1.3.2
packaging==23.2
pillow==10.2.0
pluggy==1.3.0
prompt-toolkit==3.0.43
pycodestyle==2.11.1
pycparser==2.21
pyflakes==3.2.0
PyJWT==2.8.0
pytest==7.4.4
pytest-django==4.7.0
pytest-mock==3.12.0
python-crontab==3.0.0
python-dateutil==2.8.2
python-decouple==3.8
python3-openid==3.2.0
pytz==2023.3.post1
PyYAML==6.0.1
redis==5.0.1
referencing==0.32.1
requests==2.31.0
requests-oauthlib==1.3.1
rpds-py==0.17.1
simplejson==3.19.2
six==1.16.0
sqlparse==0.4.4
tzdata==2023.4
uritemplate==4.1.1
urllib3==2.1.0
vine==5.1.0
wcwidth==0.2.13
